% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_drivers.R
\name{dbc_list}
\alias{dbc_list}
\title{List installed ADBC drivers}
\usage{
dbc_list()
}
\value{
A data frame with columns \code{id}, \code{name}, \code{version},
  \code{level} (\code{"user"}, \code{"system"}, or \code{"env"}), and
  \code{path}.
}
\description{
Equivalent to running \code{dbc list} on the command line.
}
